requirements.txt
marshmallow
python-dotenv
flask
flask_sqlalchemy
flask_migrate
pytest